+++ packages/language-css/src/expressions/__tests__/values.ts
@@ -28,6 +28,7 @@ const comments: string[] = [
   "/**/",
   "/* */",
   "/* foobar */",
+  "/* .foo{content:'bar';} */",
 ];
 
 /**